feat: Hide Market related links (in Utilities) and Routes for un-marketed networks + Refactor network utils functions to own helper file
