+++ client/src/app/components/header/Header.tsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 import { ReactComponent as Logo } from "../../../assets/logo-header.svg";
 import { ReactComponent as ShimmerLogo } from "../../../assets/shimmer-logo-header.svg";
 import { ServiceFactory } from "../../../factories/serviceFactory";
-import { ALPHANET, CUSTOM, isShimmerNetwork, TESTNET } from "../../../models/config/networkType";
+import { isMarketedNetwork, isShimmerNetwork } from "../../../helpers/networkHelper";
 import { SettingsService } from "../../../services/settingsService";
 import FiatSelector from "../FiatSelector";
 import "./Header.scss";
@@ -55,9 +55,7 @@ class Header extends Component<HeaderProps, HeaderState> {
     public render(): ReactNode {
         const { rootPath, currentNetwork, networks, history, action, search, utilities, pages } = this.props;
         const isShimmer = isShimmerNetwork(currentNetwork?.network);
-        const isMarketedNetwork = currentNetwork?.network !== ALPHANET &&
-            currentNetwork?.network !== TESTNET &&
-            currentNetwork?.network !== CUSTOM;
+        const isMarketed = isMarketedNetwork(currentNetwork?.network);
 
         return (
             <header className={classNames({ "shimmer-header-bg": isShimmer })}>
@@ -146,7 +144,7 @@ class Header extends Component<HeaderProps, HeaderState> {
                                 )}
                             </div>
                             {/* ----- Only visible in mobile ----- */}
-                            {isMarketedNetwork && (
+                            {isMarketed && (
                                 <div className="mobile-fiat">
                                     <FiatSelector />
                                 </div>
@@ -156,7 +154,7 @@ class Header extends Component<HeaderProps, HeaderState> {
                             {search}
 
                             {/* ----- Only visible in desktop ----- */}
-                            {isMarketedNetwork && (
+                            {isMarketed && (
                                 <div className="desktop-fiat">
                                     <FiatSelector />
                                 </div>
